requirements.txt
fastapi
uvicorn
pydantic
langchain-openai
langchain-pinecone
langgraph
python-dotenv